LPD-82332 Poshi to Playwright: Migrating tests of objectFieldsDateTime.spec.ts
